% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{gm_set}
\alias{gm_set}
\title{Methylation data from TCGA mapped to genomic location}
\format{An \code{\link[minfi]{GenomicMethylSet}} object}
\usage{
gm_set
}
\value{
A \code{\link[minfi]{GenomicMethylSet}} object with TCGA methylation data
}
\description{
See vignette `data_preprocessing` for code to build this object.
}
\keyword{datasets}
