% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{gratio_set}
\alias{gratio_set}
\title{Methylation data from TCGA mapped to genomic location in ratio format}
\format{A \code{\link[minfi]{GenomicRatioSet}} object}
\usage{
gratio_set
}
\value{
A \code{\link[minfi]{GenomicRatioSet}} object with TCGA methylation data
}
\description{
See vignette `data_preprocessing` for code to build this object.
}
\keyword{datasets}

