% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vignette_data.R
\docType{data}
\name{meth_set}
\alias{meth_set}
\title{Methylation data from TCGA}
\format{A \code{\link[minfi]{MethylSet}} object}
\usage{
meth_set
}
\value{
A \code{\link[minfi]{MethylSet}} object
}
\description{
See vignette `data_preprocessing` for code to build this object.
}
\keyword{datasets}

